Traces fixes
